requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.45.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-social-share==2.2.1
djangorestframework==3.12.4
drf-jwt==1.19.1
gunicorn==20.1.0
idna==3.2
oauthlib==3.1.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
whitenoise==5.2.0
